+++ beaker/data_model/group.py
@@ -1,5 +1,6 @@
 from datetime import datetime
-from typing import Optional
+from enum import Enum
+from typing import List, Optional
 
 from .account import Account
 from .base import BaseModel
@@ -15,3 +16,28 @@ class Group(BaseModel):
     created: datetime
     modified: datetime
     workspace_ref: Optional[WorkspaceRef] = None
+
+
+class GroupSpec(BaseModel):
+    workspace: Optional[str] = None
+    name: Optional[str] = None
+    description: Optional[str] = None
+    experiments: Optional[List[str]] = None
+
+
+class GroupParameterType(str, Enum):
+    metric = "metric"
+    env = "env"
+
+
+class GroupParameter(BaseModel):
+    type: GroupParameterType
+    name: str
+
+
+class GroupPatch(BaseModel):
+    name: Optional[str] = None
+    description: Optional[str] = None
+    add_experiments: Optional[List[str]] = None
+    remove_experiments: Optional[List[str]] = None
+    parameters: Optional[List[GroupParameter]] = None