+++ beaker/data_model/workspace.py
@@ -1,5 +1,6 @@
 from datetime import datetime
-from typing import Dict, Optional, Tuple
+from enum import Enum
+from typing import Dict, List, Optional, Tuple
 
 from .account import Account
 from .base import BaseModel
@@ -35,10 +36,43 @@ class WorkspacePage(BaseModel):
     next_cursor: Optional[str] = None
 
 
+class WorkspaceSpec(BaseModel):
+    name: Optional[str] = None
+    description: Optional[str] = None
+    public: bool = False
+    org: Optional[str] = None
+
+
+class WorkspaceTransferSpec(BaseModel):
+    ids: List[str]
+
+
+class Permission(str, Enum):
+    """
+    Workspace permission levels.
+    """
+
+    no_permission = "none"
+    read = "read"
+    write = "write"
+    full_control = "all"
+
+
 class WorkspacePermissions(BaseModel):
     requester_auth: str
     public: bool
-    authorizations: Dict[str, str]
+    authorizations: Dict[str, Permission]
     """
     A dictionary of account IDs to authorizations.
     """
+
+
+class WorkspacePatch(BaseModel):
+    name: Optional[str] = None
+    description: Optional[str] = None
+    archive: Optional[bool] = None
+
+
+class WorkspacePermissionsPatch(BaseModel):
+    public: Optional[bool] = None
+    authorizations: Optional[Dict[str, Permission]] = None