+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `description` parameter to `Beaker.dataset.create()`, `Beaker.image.create()`.
+- Added `description` parameter to `Beaker.dataset.create()`, `Beaker.image.create()`, `Beaker.workspace.create()`.
 
 ## [v0.13.2](https://github.com/allenai/beaker-py/releases/tag/v0.13.2) - 2022-04-29